% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Square function
edit_r_environ()
Takes a vector and returns a vector of squared components.}
\usage{
myf(x)
}
\arguments{
\item{x}{}
}
\value{
a vector of squared components
}
\description{
Square function
edit_r_environ()
Takes a vector and returns a vector of squared components.
}
\examples{
x=1:30; myf(x)
}
